answearing to the question number 5 designconsideration
i implemented the code for making the check on the critical speed , and i tried to fix the problem of critical speed working on gyroscopic effect. In the end i create a code that make a new check to be sure that no critical speed is in the non acceptable range.